- implement evaluation on test set. - implement cmd to run the same evaluation per run according to the paper - implement maker of table summary metrics

+++ model.py
@@ -43,7 +43,7 @@ class PrototypicalNetwork(nn.Module):
         tg.guard(embeddings_supp, "*, SUPP_SIZE, NUM_FEATURES")
         tg.guard(embeddings_query, "*, QUERY_SIZE, NUM_FEATURES")
         y_supp_broadcast = y_supp.unsqueeze(-1).expand(bs, supp_size, embeddings_supp.shape[-1])
-        centroids = torch.zeros(bs, num_classes, embeddings_supp.shape[-1], device=X_supp.device)\
+        centroids = torch.zeros(bs, num_classes, embeddings_supp.shape[-1], device=X_supp.device, dtype=embeddings_supp.dtype)\
                     .scatter_add(1, y_supp_broadcast, embeddings_supp) / supp_size * num_classes
         tg.guard(centroids, "*, NUM_CLASSES, NUM_FEATURES")
         result = dict(centroids=centroids,
